+++ pom.xml
@@ -3,9 +3,9 @@
                       http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.brage</groupId>
+    <groupId>ro.brage</groupId>
     <artifactId>dodo</artifactId>
-    <version>1.1.0-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>dodo</name>
 
@@ -111,22 +111,18 @@
         <developerConnection>scm:git:git@github.com:nirodg/dodo.git</developerConnection>
     </scm>
 
-    <repositories>
-        <repository>
-            <id>myMavenRepoRead</id>
-            <url>${myMavenRepoReadUrl}</url>
-        </repository>
-    </repositories>
 
-    <distributionManagement>
-        <repository>
-            <id>myMavenRepoWrite</id>
-            <url>${myMavenRepoWriteUrl}</url>
-        </repository>
-        <snapshotRepository>
-            <id>myMavenRepoWrite</id>
-            <url>${myMavenRepoWriteUrl}</url>
-        </snapshotRepository>
-    </distributionManagement>
+	<distributionManagement>
+		<repository>
+		  <id>nexus</id>
+		  <name>Releases</name>
+		  <url>http://eorghe.com:8081/repository/maven-releases</url>
+		</repository>
+		<snapshotRepository>
+		  <id>nexus</id>
+		  <name>Snapshot</name>
+		  <url>http://eorghe.com:8081/repository/maven-snapshots</url>
+		</snapshotRepository>
+	</distributionManagement>
 
 </project>
