+++ src/main/java/ro/brage/dodo/jpa/enums/OrderBy.java
@@ -1,4 +1,4 @@
-package com.brage.dodo.jpa.enums;
+package ro.brage.dodo.jpa.enums;
 
 public enum OrderBy {
   ASC, DESC;
