+++ src/main/java/ro/brage/dodo/jpa/mapper/AbstractModelMapper.java
@@ -16,63 +16,40 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package com.brage.dodo.jpa;
+package ro.brage.dodo.jpa.mapper;
 
-import java.io.Serializable;
-import java.util.Date;
+import java.util.List;
+import org.mapstruct.MapperConfig;
+import org.mapstruct.MappingTarget;
+import ro.brage.dodo.jpa.AbstractDTOModel;
+import ro.brage.dodo.jpa.Model;
+import ro.brage.dodo.jpa.mapper.qualifiers.LoadEntity;
 
 /**
  *
  * @author Dorin Brage
+ * @param <Entity>
+ * @param <DTO>
  */
-public class AbstractDTOModel implements Serializable {
+@MapperConfig(componentModel = "cdi")
+public abstract class AbstractModelMapper<Entity extends Model, DTO extends AbstractDTOModel> {
 
-  private static final long serialVersionUID = -4361997507068841444L;
+  public abstract Entity find(DTO entity);
 
-  private String guid;
-  private String createdBy;
-  private String updatedBy;
-  private Date createdOn;
-  private Date updatedOn;
+  public abstract DTO find(Entity entity);
 
-  public String getGuid() {
-    return guid;
-  }
+  public abstract void updateEntity(DTO dto, @MappingTarget Entity entity);
 
-  public void setGuid(String guid) {
-    this.guid = guid;
-  }
+  public abstract void updateDTO(@MappingTarget DTO dto, Entity entity);
 
-  public String getCreatedBy() {
-    return createdBy;
-  }
+  public abstract List<DTO> findDTOs(List<Entity> entities);
 
-  public void setCreatedBy(String createdBy) {
-    this.createdBy = createdBy;
-  }
+  public abstract List<Entity> findEntities(List<DTO> dtos);
 
-  public String getUpdatedBy() {
-    return updatedBy;
-  }
+  @LoadEntity
+  public abstract Entity load(DTO dto);
 
-  public void setUpdatedBy(String updatedBy) {
-    this.updatedBy = updatedBy;
-  }
-
-  public Date getCreatedOn() {
-    return createdOn;
-  }
-
-  public void setCreatedOn(Date createdOn) {
-    this.createdOn = createdOn;
-  }
-
-  public Date getUpdatedOn() {
-    return updatedOn;
-  }
-
-  public void setUpdatedOn(Date updatedOn) {
-    this.updatedOn = updatedOn;
-  }
+  @LoadEntity
+  public abstract DTO load(Entity entity);
 
 }
