+++ src/main/java/ro/brage/dodo/rs/AbstractRestService.java
@@ -16,16 +16,46 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package com.brage.dodo.jpa.enums;
+package ro.brage.dodo.rs;
+
+import java.util.List;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.DELETE;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+import ro.brage.dodo.jpa.AbstractDTOModel;
 
 /**
  *
  * @author Dorin Brage
  */
-public enum JpaErrorKeys {
-  // FIND
-  FAILED_TO_FIND_ENTITY, FAILED_TO_FIND_ENTITIES,
+@Consumes(MediaType.APPLICATION_JSON)
+@Produces(MediaType.APPLICATION_JSON)
+public interface AbstractRestService<DTO extends AbstractDTOModel> {
 
-  // LOAD
-  FAILED_TO_LOAD_ENTITY, FAILED_TO_LOAD_ENTITIES,
+  @GET
+  @Path("/")
+  public List<DTO> getAll();
+
+  @POST
+  @Path("/")
+  public DTO create(DTO entity);
+
+  @PUT
+  @Path("/{guid}")
+  public DTO updateByGuid(@PathParam("guid") String guid, DTO entity);
+
+  @GET
+  @Path("/{guid}")
+  public DTO getByGuid(@PathParam("guid") String guid);
+
+  @DELETE
+  @Path("/{guid}")
+  public boolean deleteByGuid(@PathParam("guid") String guid);
+
 }
